requirements.txt
APScheduler>=3.10.4
PyYAML>=6.0.1
requests>=2.32.0
feedparser>=6.0.11
typer>=0.12.3
streamlit>=1.36.0
pytest>=8.2.0
responses>=0.25.0
